refactor: group sidebar panel modules
