+++ src/webview/panels/integrityView.ts
@@ -1,1 +1,1 @@
-export * from './integrity/integrityView';
+export * from './integrity/index';
