+++ src/webview/panels/struct/index.ts
